fix(neuron): cast neuron constants to uint type consistently
Resolve type mismatch issues by ensuring all neuron constants
are properly cast to uint type when used in array initializations
and function calls across multiple neuron cell implementations.

+++ pkg/neuron/cell/bias.go
@@ -12,7 +12,7 @@ type Bias[T utils.Float] core[T]
 
 func _NewBias[T utils.Float]() *Bias[T] {
 	return &Bias[T]{
-		Id:    [2]uint{neuron.BIAS, 0},
+		Id:    [2]uint{uint(neuron.BIAS), 0},
 		value: 1.0,
 	}
 }
